Enhanced Choosing Sim in CellManager and Added setParameters method to Rules

+++ CellSociety/src/Cell.java
@@ -2,8 +2,8 @@
 public class Cell {
     private int currState = 0;
     private int nextState = 0;
-    private int turnsOnState = 0;
-    private int energy;
+    private double turnsOnState = 0;
+    private double energy;
 
 
     public int getCurrState () {
@@ -23,18 +23,18 @@ public class Cell {
     }
     
     
-    public int getTurnsOnState () {
+    public double getTurnsOnState () {
         return turnsOnState;
     }
 
-    public void setTurnsOnState (int turnsOnState) {
+    public void setTurnsOnState (double turnsOnState) {
         this.turnsOnState = turnsOnState;
     }
-    public int getEnergy () {
+    public double getEnergy () {
         return energy;
     }
 
-    public void setEnergy (int energy) {
+    public void setEnergy (double energy) {
         this.energy = energy;
     }
